+++ addon/components/object-list-view.js
@@ -14,6 +14,16 @@ import { translationMacro as t } from 'ember-i18n';
  * @extends FlexberryBaseComponent
  */
 export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentMixin, ErrorableMixin, {
+  /**
+    Projection set by property `modelProjection`.
+
+    @property _modelProjection
+    @type Object
+    @default null
+    @private
+   */
+  _modelProjection: null,
+
   actions: {
     rowClick: function(recordWithKey, e) {
       if (this.get('readonly')) {
@@ -357,9 +367,9 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
     'modelProjection',
     function() {
     if (this.get('modelProjection')) {
-      return false;
-    } else {
       return this.get('showAsteriskInRow') || this.get('showCheckBoxInRow') || this.get('showDeleteButtonInRow');
+    } else {
+      return false;
     }
   }),
 
@@ -440,13 +450,34 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
   ),
 
   /**
-   * Model projection which should be used to display given content.
-   *
-   * @property modelProjection
-   * @type Object
-   * @default null
+    Model projection which should be used to display given content.
+    Accepts object or name projections.
+
+    @property modelProjection
+    @type Object|String
+    @default null
    */
-  modelProjection: null,
+  modelProjection: Ember.computed('_modelProjection', {
+    get(key) {
+      return this.get('_modelProjection');
+    },
+    set(key, value) {
+      if (typeof value === 'string') {
+        let modelName = this.get('modelName');
+        Ember.assert('For define projection by name, model name is required.', modelName);
+        let modelConstructor = this.get('store').modelFor(modelName);
+        Ember.assert(`Model with name '${modelName}' is not found.`, modelConstructor);
+        let projections = Ember.get(modelConstructor, 'projections');
+        Ember.assert(`Projection with name '${value}' for model with name '${modelName}' is not found.`, projections[value]);
+        value = projections[value];
+      } else if (typeof value !== 'object') {
+        throw new Error(`Property 'modelProjection' should be a string or object.`);
+      }
+
+      this.set('_modelProjection', value);
+      return value;
+    },
+  }),
 
   /**
    * Table columns related to current model projection.
@@ -822,10 +853,19 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
     this.get('objectlistviewEventsService').on('olvDeleteRows', this, this._deleteRows);
     this.get('objectlistviewEventsService').on('filterByAnyMatch', this, this._filterByAnyMatch);
 
-    if (this.get('content')) {
-      this.get('content').forEach((item, index, enumerable) => {
-        this._addModel(item);
-      });
+    let content = this.get('content');
+    if (content) {
+      if (content.get('isFulfilled') === false) {
+        content.then((items) => {
+          items.forEach((item) => {
+            this._addModel(item);
+          });
+        });
+      } else {
+        content.forEach((item) => {
+          this._addModel(item);
+        });
+      }
     }
   },
 
@@ -869,6 +909,11 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
     getSettingPromise.then(function(data) {
       _this._setColumnWidths(data);
     });
+
+    // TODO: resolv this problem.
+    this.$('.flexberry-dropdown:last').dropdown({
+      direction: 'upward'
+    });
   },
 
   /**
@@ -877,8 +922,11 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
    *
    * @method didRender
    */
-  didRender: function() {
+  didRender() {
     this._super(...arguments);
+
+    let currentTable = this.$('table.object-list-view');
+
     if (this.get('allowColumnResize')) {
       if (this.get('useSingleColumn')) {
         Ember.Logger.error(
@@ -886,8 +934,6 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
           'can\'t be enabled at the same time.');
         return;
       }
-
-      let currentTable = this.$('table.object-list-view');
 
       // The first column has semantic class "collapsing"
       // so the column has 1px width and plugin has problems.
@@ -900,6 +946,8 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
       currentTable.addClass('fixed');
 
       this._reinitResizablePlugin();
+    } else {
+      currentTable.colResizable({ disable: true });
     }
   },
 
@@ -910,18 +958,18 @@ export default FlexberryBaseComponent.extend(FlexberryLookupCompatibleComponentM
    * @method _reinitResizablePlugin
    * @private
    */
-  _reinitResizablePlugin: function() {
+  _reinitResizablePlugin() {
     let currentTable = this.$('table.object-list-view');
 
     // Disable plugin and then init it again.
     currentTable.colResizable({ disable: true });
 
-    let _this = this;
     currentTable.colResizable({
-      minWidth: 70,
-      onResize: function(e) {
+      minWidth: 90,
+      resizeMode:'flex',
+      onResize: (e)=> {
         // Save column width as user setting on resize.
-        _this._afterColumnResize(e);
+        this._afterColumnResize(e);
       }
     });
   },