+++ tests/dummy/app/routes/components-examples/flexberry-objectlistview/on-edit-form.js
@@ -1,18 +1,17 @@
-import EditFormRoute from 'ember-flexberry/routes/edit-form';
-import EditFormMixin from '../mixins/edit-form-mixin';
+import ListFormRoute from 'ember-flexberry/routes/list-form';
 
-export default EditFormRoute.extend(EditFormMixin, {
+export default ListFormRoute.extend({
   /**
     Name of model projection to be used as record's properties limitation.
 
     @property modelProjection
     @type String
-    @default 'ApplicationUserE'
+    @default 'ApplicationUserL'
    */
-  modelProjection: 'ApplicationUserE',
+  modelProjection: 'ApplicationUserL',
 
   /**
-    Name of model to be used as form's record type.
+    Name of model to be used as list's records types.
 
     @property modelName
     @type String