+++ tests/dummy/app/routes/components-examples/flexberry-objectlistview/on-edit-form/user.js
@@ -1,7 +1,6 @@
-import EditFormRoute from 'ember-flexberry/routes/edit-form';
-import EditFormMixin from '../mixins/edit-form-mixin';
+import BaseEditFormRoute from '../../../base-edit-form';
 
-export default EditFormRoute.extend(EditFormMixin, {
+export default BaseEditFormRoute.extend({
   /**
     Name of model projection to be used as record's properties limitation.